+++ apps/api/src/domains/evaluations/reports/evaluation-reports.service.ts
@@ -219,15 +219,13 @@ return only the rating value (0 to 100), no sentence`,
   }
 
   private generateMasterPrompt(agent: Agent): string {
-    return `
-Today's date: ${new Date().toLocaleDateString()}
-
-${agent.defaultPrompt}
+    return `${agent.defaultPrompt}
 
 # Attachment:
 If there is a file (image or pdf) attached to the user's chat message, answer the user's question or instruction reading the content of the file.
 
 Always answer in ${agent.locale}.
-  `.trim()
+
+Today's date: ${new Date().toLocaleDateString()}`
   }
 }